+++ Rdmp.Core.Tests/DataLoad/Engine/Integration/CheckingTests/ProcessTaskCheckingTests.cs
@@ -36,7 +36,10 @@ public class ProcessTaskCheckingTests : DatabaseTests
         _dir.Create();
 
         var hicdir = LoadDirectory.CreateDirectoryStructure(_dir, "ProjDir", true);
-        _lmd.LocationOfFlatFiles = hicdir.RootPath.FullName;
+        _lmd.LocationOfForLoadingDirectory = Path.Combine(hicdir.RootPath.FullName, _lmd.DefaultForLoadingPath);
+        _lmd.LocationOfForArchivingDirectory = Path.Combine(hicdir.RootPath.FullName, _lmd.DefaultForArchivingPath);
+        _lmd.LocationOfExecutablesDirectory = Path.Combine(hicdir.RootPath.FullName, _lmd.DefaultExecutablesPath);
+        _lmd.LocationOfCacheDirectory = Path.Combine(hicdir.RootPath.FullName, _lmd.DefaultCachePath);
         _lmd.SaveToDatabase();
 
         var c = new Catalogue(CatalogueRepository, "c");
@@ -104,7 +107,10 @@ public class ProcessTaskCheckingTests : DatabaseTests
     [Test]
     public void MEFCompatibleType_NoProjectDirectory()
     {
-        _lmd.LocationOfFlatFiles = null;
+        _lmd.LocationOfForLoadingDirectory = null;
+        _lmd.LocationOfForArchivingDirectory = null;
+        _lmd.LocationOfExecutablesDirectory = null;
+        _lmd.LocationOfCacheDirectory = null;
         _lmd.SaveToDatabase();
 
         _task.ProcessTaskType = ProcessTaskType.Attacher;
@@ -114,7 +120,7 @@ public class ProcessTaskCheckingTests : DatabaseTests
         _task.CreateArgumentsForClassIfNotExists<AnySeparatorFileAttacher>();
 
         var ex = Assert.Throws<Exception>(() => _checker.Check(ThrowImmediatelyCheckNotifier.QuietPicky));
-        Assert.That(ex.InnerException.Message, Is.EqualTo($@"No Project Directory (LocationOfFlatFiles) has been configured on LoadMetadata {_lmd.Name}"));
+        Assert.That(ex.InnerException.Message, Is.EqualTo($@"No Project Directory has been configured on LoadMetadata {_lmd.Name}"));
     }
 
     [Test]
@@ -125,7 +131,10 @@ public class ProcessTaskCheckingTests : DatabaseTests
                 "DelMeProjDir", true);
         try
         {
-            _lmd.LocationOfFlatFiles = projDir.RootPath.FullName;
+            _lmd.LocationOfForLoadingDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultForLoadingPath);
+            _lmd.LocationOfForArchivingDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultForArchivingPath);
+            _lmd.LocationOfExecutablesDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultExecutablesPath);
+            _lmd.LocationOfCacheDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultCachePath);
             _task.ProcessTaskType = ProcessTaskType.Attacher;
             _task.LoadStage = LoadStage.Mounting;
             _task.Path = typeof(AnySeparatorFileAttacher).FullName;
@@ -152,7 +161,10 @@ public class ProcessTaskCheckingTests : DatabaseTests
                 "DelMeProjDir", true);
         try
         {
-            _lmd.LocationOfFlatFiles = projDir.RootPath.FullName;
+            _lmd.LocationOfForLoadingDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultForLoadingPath);
+            _lmd.LocationOfForArchivingDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultForArchivingPath);
+            _lmd.LocationOfExecutablesDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultExecutablesPath);
+            _lmd.LocationOfCacheDirectory = Path.Combine(projDir.RootPath.FullName, _lmd.DefaultCachePath);
             _task.ProcessTaskType = ProcessTaskType.Attacher;
             _task.LoadStage = LoadStage.Mounting;
             _task.Path = typeof(AnySeparatorFileAttacher).FullName;