+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandAddCohortToExtractionConfiguration.cs
@@ -58,12 +58,9 @@ public class ExecuteCommandAddCohortToExtractionConfiguration : BasicCommandExec
 
         if (_targetExtractionConfiguration.Cohort_ID != null)
         {
-            if (_targetExtractionConfiguration.Cohort_ID == sourceExtractableCohortComand.Cohort.ID)
-                SetImpossible("ExtractionConfiguration already uses this cohort");
-            else
-                SetImpossible(
-                    "ExtractionConfiguration already uses a different cohort (delete the relationship to the old cohort first)");
-
+            SetImpossible(_targetExtractionConfiguration.Cohort_ID == sourceExtractableCohortComand.Cohort.ID
+                ? "ExtractionConfiguration already uses this cohort"
+                : "ExtractionConfiguration already uses a different cohort (delete the relationship to the old cohort first)");
             return;
         }
     }